XWIKI-21670: The versionPreserved instance output filter stream property defaults to false
* Deduplicate common properties from DocumentInstanceOutputProperties and
  UserInstanceOutputProperties in a common super class.
