+++ xwiki-platform-core/xwiki-platform-oldcore/src/main/java/com/xpn/xwiki/internal/filter/output/UserInstanceOutputProperties.java
@@ -19,8 +19,7 @@
  */
 package com.xpn.xwiki.internal.filter.output;
 
-import org.xwiki.filter.instance.output.InstanceOutputProperties;
-import org.xwiki.model.reference.DocumentReference;
+import org.xwiki.filter.instance.output.EntityInstanceOutputProperties;
 import org.xwiki.properties.annotation.PropertyDescription;
 import org.xwiki.properties.annotation.PropertyName;
 
@@ -28,28 +27,8 @@ import org.xwiki.properties.annotation.PropertyName;
  * @version $Id$
  * @since 6.2M1
  */
-public class UserInstanceOutputProperties extends InstanceOutputProperties
+public class UserInstanceOutputProperties extends EntityInstanceOutputProperties
 {
-    /**
-     * @see #getSaveComment()
-     */
-    private String saveComment = "Import";
-
-    /**
-     * @see #getAuthor()
-     */
-    private DocumentReference author;
-
-    /**
-     * @see #isAuthorSet()
-     */
-    private boolean authorSet;
-
-    /**
-     * @see #isVersionPreserved()
-     */
-    private boolean versionPreserved = true;
-
     /**
      * @see #getGroupPrefix()
      */
@@ -59,71 +38,6 @@ public class UserInstanceOutputProperties extends InstanceOutputProperties
      * @see #getGroupSuffix()
      */
     private String groupSuffix = "";
-
-    /**
-     * @return the comment to use when saving users
-     */
-    @PropertyName("Save comment")
-    @PropertyDescription("The comment to use when saving a user")
-    public String getSaveComment()
-    {
-        return this.saveComment;
-    }
-
-    /**
-     * @param saveComment the version comment to use when saving users
-     */
-    public void setSaveComment(String saveComment)
-    {
-        this.saveComment = saveComment;
-    }
-
-    /**
-     * @return the author to use when saving users
-     */
-    @PropertyName("Save author")
-    @PropertyDescription("The author to use when saving a user")
-    public DocumentReference getAuthor()
-    {
-        return this.author;
-    }
-
-    /**
-     * @param author the author to use when saving users
-     */
-    public void setAuthor(DocumentReference author)
-    {
-        this.author = author;
-        this.authorSet = true;
-    }
-
-    /**
-     * @return true if the author have been explicitly set
-     */
-    public boolean isAuthorSet()
-    {
-        return this.authorSet;
-    }
-
-    /**
-     * @return Indicate if the versions related informations coming from the events should be kept
-     * @since 6.2M1
-     */
-    @PropertyName("Preserve version")
-    @PropertyDescription("Indicate if the versions related informations coming from the events should be kept")
-    public boolean isVersionPreserved()
-    {
-        return this.versionPreserved;
-    }
-
-    /**
-     * @param versionPreserved Indicate if the versions related informations coming from the events should be kept
-     * @since 6.2M1
-     */
-    public void setVersionPreserved(boolean versionPreserved)
-    {
-        this.versionPreserved = versionPreserved;
-    }
 
     /**
      * @return the prefix to add before each new group name
